Rewrite to handle all data blocks instead of just textures
